requirements.txt
asgiref==3.9.2
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
google-auth==2.41.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
